# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Velora CRM API Documentation",
    "version": "1.2.0",
    "description": "\n# WhatsApp AI Gateway - Complete API Reference\n\nProfessional WhatsApp Gateway with **64 API endpoints** for complete WhatsApp automation.\n\n## 🔐 Authentication\nAll endpoints require authentication via:\n1. **API Key Header**: `X-API-Key: your-api-key`\n2. **Session Cookie**: `next-auth.session-token` (automatically sent by browser)\n\n## 📋 Common Parameters\n- **sessionId**: Unique session identifier (e.g., \"mysession-01\")\n- **jid**: WhatsApp JID format:\n  - Personal: `628123456789@s.whatsapp.net`\n  - Group: `120363123456789@g.us`\n\n## 📊 Rate Limits\n- Phone check: Max 50 numbers per request\n- Broadcast: 10-20s random delay between messages\n- Message history: Max 100 messages\n                "
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key for authentication"
      },
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "Session cookie (browser only)"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Error occurred"
          },
          "error": {
            "type": "string",
            "example": "Detailed error info"
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Unauthorized - Invalid or missing API key",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "status": false,
              "message": "Unauthorized",
              "error": "Unauthorized"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden - Access denied",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "status": false,
              "message": "Forbidden - Cannot access this session",
              "error": "Forbidden - Cannot access this session"
            }
          }
        }
      },
...
      "SessionNotReady": {
        "description": "Session not connected or ready",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "status": false,
              "message": "Session not ready",
              "error": "Session not ready"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "SessionAuth": []
    }
  ],
  "paths": {
...
    "/messages/{sessionId}/{jid}/sticker": {
      "post": {
        "tags": [
          "Messaging"
        ],
        "summary": "Send sticker",
        "description": "Convert an image to sticker and send it",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "jid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "pack": {
                    "type": "string",
                    "description": "Sticker pack name (default: Velora CRM)"
                  },
                  "author": {
                    "type": "string",
                    "description": "Sticker author name (default: User)"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "full",
                      "crop",
                      "circle"
                    ],
                    "description": "Sticker crop type (default: full)"
                  },
                  "quality": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 100,
                    "description": "Image quality (default: 50)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sticker sent",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "description": "Failed to create sticker"
          },
          "503": {
            "$ref": "#/components/responses/SessionNotReady"
          }
        }
      }
    },
...
    "/settings/system": {
      "post": {
        "tags": [
          "System"
        ],
        "summary": "Update system settings",
        "description": "Update global system configuration (Superadmin/Owner only)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "appName": {
                    "type": "string"
                  },
                  "logoUrl": {
                    "type": "string"
                  },
                  "timezone": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Settings updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "appName": {
                      "type": "string"
                    },
                    "logoUrl": {
                      "type": "string"
                    },
                    "timezone": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "description": "Failed to update settings"
          }
        }
      },
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Get system settings",
        "responses": {
          "200": {
            "description": "System settings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "default"
                    },
                    "appName": {
                      "type": "string",
                      "example": "Velora CRM"
                    },
                    "logoUrl": {
                      "type": "string",
                      "example": "https://example.com/logo.png"
                    },
                    "timezone": {
                      "type": "string",
                      "example": "Asia/Jakarta"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}